requirements.txt
#由Gemini-2.5-Pro生成
# --- 核心框架 ---
nonebot2  # NoneBot2 核心框架
nonebot-adapter-onebot  # OneBot V11 适配器

# --- 网络与API请求 ---
httpx  # 高性能的HTTP客户端，用于调用API

# --- 配置与数据处理 ---
python-dotenv  # 用于从 .env 文件加载环境变量
pydantic  # 用于数据验证和模型定义 (data_store.py 中大量使用)
PyYAML  # 用于解析 jm_option.yml 配置文件

# --- 功能性依赖 ---
jmcomic  # 禁漫下载核心库
duckduckgo-search  # DDGS 搜索库，用于网页和新闻搜索工具